+++ test/format.js
@@ -22,7 +22,10 @@ describe('format', () =>
 	{
 		// Switzerland
 		format({ country: 'CH', phone: '446681800' }, 'International').should.equal('+41 44 668 18 00')
+		format({ country: 'CH', phone: '446681800' }, 'E.164').should.equal('+41446681800')
+		// "International_plaintext" is deprecated
 		format({ country: 'CH', phone: '446681800' }, 'International_plaintext').should.equal('+41446681800')
+		format({ country: 'CH', phone: '446681800' }, 'RFC3966').should.equal('+41446681800')
 		format({ country: 'CH', phone: '446681800' }, 'National').should.equal('044 668 18 00')
 
 		// France
@@ -72,5 +75,53 @@ describe('format', () =>
 	{
 		local_to_international_style('(xxx) xxx-xx-xx').should.equal('xxx xxx xx xx')
 		local_to_international_style('(xxx)xxx').should.equal('xxx xxx')
+	})
+
+	it('should format phone number extensions', function()
+	{
+		// National
+		format
+		({
+			country: 'US',
+			phone: '2133734253',
+			ext: '123'
+		},
+		'National').should.equal('(213) 373-4253 ext. 123')
+
+		// International
+		format
+		({
+			country : 'US',
+			phone   : '2133734253',
+			ext     : '123'
+		},
+		'International').should.equal('+1 213 373 4253 ext. 123')
+
+		// International
+		format
+		({
+			country : 'US',
+			phone   : '2133734253',
+			ext     : '123'
+		},
+		'International').should.equal('+1 213 373 4253 ext. 123')
+
+		// E.164
+		format
+		({
+			country : 'US',
+			phone   : '2133734253',
+			ext     : '123'
+		},
+		'E.164').should.equal('+12133734253')
+
+		// RFC3966
+		format
+		({
+			country : 'US',
+			phone   : '2133734253',
+			ext     : '123'
+		},
+		'RFC3966').should.equal('+12133734253;ext=123')
 	})
 })